wxToString() doesn't support "unsigned long long", so size_t doesn't work on Windows x64 builds.

+++ src/effects/Equalization.cpp
@@ -314,7 +314,7 @@ EffectType EffectEqualization::GetType()
 
 bool EffectEqualization::GetAutomationParameters(EffectAutomationParameters & parms)
 {
-   parms.Write(KEY_FilterLength, mM);
+   parms.Write(KEY_FilterLength, (unsigned long)mM);
    parms.Write(KEY_CurveName, mCurveName);
    parms.Write(KEY_InterpLin, mLin);
    parms.WriteEnum(KEY_InterpMeth, mInterp, wxArrayString(kNumInterpolations, kInterpStrings));
